fixed bug when taking acos Original: 2efc97404a83fcfa176dda490e82234fd4211a49

+++ lalapps/src/pulsar/FDS_isolated/polka.c
@@ -7,6 +7,7 @@
 /*                                  UWM - March  2004                            */
 /*********************************************************************************/
 
+
 #include <unistd.h>
 #include <stdio.h>
 #include <string.h>
@@ -19,7 +20,7 @@
 #include <lal/LALConstants.h>
 
 #define MAXCANDIDATES   3750000     /* Maximum # of allowed candidates */
-#define MAXCOINC   3750000          /* Maximum # of allowed coincident candidates */
+#define MAXCOINC   3750000         /* Maximum # of allowed coincident candidates */
 
 struct CommandLineArgsTag 
 {
@@ -166,7 +167,7 @@ int main(int argc,char *argv[])
 	  n2[1]=sin(Alpha2)*cos(Delta2);
 	  n2[2]=sin(Delta2);
 
-	  AngularDistance=acos(n1[0]*n2[0]+n1[1]*n2[1]+n1[2]*n2[2]);
+	  AngularDistance=acos((float)(n1[0]*n2[0]+n1[1]*n2[1]+n1[2]*n2[2]));
 	  difff=fabs(f1 - C2.f[indices2f[f]]);
 
 	  /* check difference in frequencies because we're not guaranteed
